Remove `ProseMirrorMenuPF2e` subclass and patch some `ProseMirrorMenu` methdos instead
