+++ src/module/item/base/sheet/sheet.ts
@@ -11,7 +11,6 @@ import {
     TraitTagifyEntry,
 } from "@module/sheet/helpers.ts";
 import { InlineRollLinks } from "@scripts/ui/inline-roll-links.ts";
-import { ProseMirrorMenuPF2e } from "@system/prosemirror-menu.ts";
 import {
     BasicConstructorOptions,
     LanguageSelector,
@@ -305,7 +304,7 @@ class ItemSheetPF2e<TItem extends ItemPF2e> extends ItemSheet<TItem, ItemSheetOp
         options: { remove?: boolean },
     ): Record<string, ProseMirror.Plugin> {
         const plugins = super._configureProseMirrorPlugins(name, options);
-        plugins.menu = ProseMirrorMenuPF2e.build(foundry.prosemirror.defaultSchema, {
+        plugins.menu = foundry.prosemirror.ProseMirrorMenu.build(foundry.prosemirror.defaultSchema, {
             destroyOnSave: options.remove,
             onSave: () => this.saveEditor(name, options),
             compact: this.options.hasSidebar,
